Making a predefined tilemap object in the MainScene
